+++ src/py42/clients/__init__.py
@@ -10,5 +10,6 @@ Clients = namedtuple(
         "securitydata",
         "auditlogs",
         "cases",
+        "loginconfig",
     ],
 )