+++ tests/test_exceptions.py
@@ -5,7 +5,6 @@ from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42ForbiddenError
 from py42.exceptions import Py42HTTPError
 from py42.exceptions import Py42InternalServerError
-from py42.exceptions import Py42MFARequiredError
 from py42.exceptions import Py42NotFoundError
 from py42.exceptions import Py42ResponseError
 from py42.exceptions import Py42TooManyRequestsError
@@ -22,18 +21,6 @@ class TestPy42Errors:
     def test_raise_py42_error_raises_unauthorized_error(self, error_response):
         error_response.response.status_code = 401
         with pytest.raises(Py42UnauthorizedError, match=REQUEST_EXCEPTION_MESSAGE):
-            raise_py42_error(error_response)
-
-    def test_raise_py42_error_raises_MFA_required_error(self, error_response):
-        error_response.response.status_code = 401
-        error_response.response.text = (
-            '{"error":[{"primaryErrorKey":"TIME_BASED_ONE_TIME_PASSWORD_REQUIRED"}]}'
-        )
-        with pytest.raises(Py42MFARequiredError):
-            raise_py42_error(error_response)
-
-        error_response.response.text = '{"error":[{"primaryErrorKey":"TOTP_AUTH_CONFIGURATION_REQUIRED_FOR_USER"}]}'
-        with pytest.raises(Py42MFARequiredError):
             raise_py42_error(error_response)
 
     def test_raise_py42_error_raises_forbidden_error(self, error_response):